+++ lumibot/components/__init__.py
@@ -1,1 +1,2 @@
 from .drift_calculation_logic import DriftCalculationLogic
+from .limit_order_drift_rebalancer_logic import LimitOrderDriftRebalancerLogic
